chore: moved getTransformedHeaders to new file
